wsl check

+++ packages/wsl/.oh-my-zsh/custom/wsl.zsh
@@ -1,5 +1,7 @@
-export BROWSER="/mnt/c/Program\ Files/Mozilla\ Firefox/firefox.exe"
+if [ $(systemd-detect-virt) = "wsl" ]; then
+    export BROWSER="/mnt/c/Program\ Files/Mozilla\ Firefox/firefox.exe"
 
-clean-zone-identifier() {
-    find . -type f -name "*:Zone.Identifier" -exec rm "{}" \;
-}
+    clean-zone-identifier() {
+        find . -type f -name "*:Zone.Identifier" -exec rm "{}" \;
+    }
+fi
